api.gamehistory/requirements.txt
flask
flask-restful
pyyaml
requests
flask-mongoengine